Rename csvfile module to datasource
