+++ src/services/oracles/ohlcoracle.mts
@@ -5,8 +5,8 @@ import type { CryptoRegistryNG } from "../../cryptoregistry.mjs";
 import { formatDate } from "../../date.mjs";
 import { Fixed } from "../../bignumber.mjs";
 import type { FixedLike } from "../../bignumber.mjs";
-import type { CSVFileOptionBag, DataSource } from "../../csvfile.mjs";
-import { CSVFile } from "../../csvfile.mjs";
+import type { CSVFileOptionBag, DataSource } from "../../datasource.mjs";
+import { CSVFile } from "../../datasource.mjs";
 import { Oracle } from "../oracle.mjs";
 import { logger } from "../../debug.mjs";
 import { GlobalMetadataStore } from "../../metadata.mjs";
